notebooks/requirements.txt
junitparser==2.0.0
dateparser
pandas
matplotlib
openpyxl
gitpython
pygount